+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/domain/MonitoringLevel.java
@@ -1,0 +1,16 @@
+package eu.bcvsolutions.idm.core.api.domain;
+
+import com.google.common.annotations.Beta;
+
+@Beta
+/**
+ * Monitoring level enum
+ *
+ * @author Vít Švanda
+ * @since 10.4.0
+ */
+public enum MonitoringLevel {
+	OK,
+	WARNING,
+	ERROR;
+}