+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/eav/api/dto/IdmFormProjectionDto.java
@@ -92,6 +92,7 @@ public class IdmFormProjectionDto
 	 * Enabled form definitions and attributes.
 	 * Json representation - [{ definition: uuid,  attributes: [ uuid, uuid ] }].
 	 * 
+	 * @see FormDefinitionAttributes
 	 * @return
 	 */
 	public String getFormDefinitions() {