+++ Realization/frontend/czechidm-core/src/components/basic/index.js
@@ -116,5 +116,5 @@ const Components = {
   ToggleSwitch
 };
 
-Components.version = '10.2.0';
+Components.version = '10.3.0';
 module.exports = Components;