+++ Realization/frontend/czechidm-core/src/content/identity/IdentityTable.js
@@ -33,7 +33,8 @@ export class IdentityTable extends Advanced.AbstractTableContent {
     super(props, context);
     this.state = {
       filterOpened: props.filterOpened,
-      showAddModal: false
+      showAddModal: false,
+      projections: []
     };
   }
 
@@ -202,6 +203,8 @@ export class IdentityTable extends Advanced.AbstractTableContent {
     const roleDisabled = _forceSearchParameters.getFilters().has('role');
     const treeNodeDisabled = _forceSearchParameters.getFilters().has('treeNodeId');
     //
+    const canCreateIdentity = showAddButton && this.getManager().canSave() && (isDefaultFormProjection || projections.length > 0);
+    //
     return (
       <Basic.Div>
         <Advanced.Table
@@ -267,7 +270,8 @@ export class IdentityTable extends Advanced.AbstractTableContent {
                     <Advanced.Filter.FormProjectionSelect
                       ref="formProjection"
                       placeholder={ this.i18n('filter.formProjection.placeholder') }
-                      manager={ projectionManager }/>
+                      manager={ projectionManager }
+                      forceSearchParameters={ new SearchParameters().setFilter('disabled', 'false') }/>
                   </Basic.Col>
                   <Basic.Col lg={ 4 }>
                     <Advanced.Filter.EnumSelectBox
@@ -302,7 +306,7 @@ export class IdentityTable extends Advanced.AbstractTableContent {
                 type="submit"
                 className="btn-xs"
                 onClick={ this.showDetail.bind(this, {}) }
-                rendered={ showAddButton && this.getManager().canSave() }
+                rendered={ canCreateIdentity }
                 icon="fa:user-plus">
                 { this.i18n('content.identity.create.button.add') }
               </Basic.Button>